+++ metadata/metadata_test.go
@@ -7,6 +7,23 @@ import (
 	report "gitlab.com/gitlab-org/security-products/analyzers/report/v3"
 )
 
+func TestReportAnalyzer(t *testing.T) {
+	want := report.AnalyzerDetails{
+		ID:      "semgrep",
+		Name:    "Semgrep",
+		Version: AnalyzerVersion,
+		Vendor: report.Vendor{
+			Name: "GitLab",
+		},
+		URL: "https://gitlab.com/gitlab-org/security-products/analyzers/semgrep",
+	}
+	got := ReportAnalyzer
+
+	if !reflect.DeepEqual(want, got) {
+		t.Errorf("Wrong result. Expected:\n%#v\nbut got:\n%#v", want, got)
+	}
+}
+
 func TestReportScanner(t *testing.T) {
 	want := report.ScannerDetails{
 		ID:      "semgrep",